mama-fresh-backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
dj-database-url==3.1.2
psycopg2-binary==2.9.11
gunicorn==25.3.0
whitenoise==6.12.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2026.1
